+++ src/index.ts
@@ -6,4 +6,10 @@ export { payload } from './utils/payload';
 
 export { InferActions } from './types/inferActions';
 
+export * from './types/common';
+export * from './types/creator';
+export * from './types/asyncCreator';
+export * from './types/creatorsPack';
+export * from './types/reducer';
+
 export { setupReducer } from './reducer';