+++ src/reducer/setupReducer.ts
@@ -75,6 +75,10 @@ export function setupReducer<State, Action extends UnknownAction<string>>(
       registerHandler(actionCreatorOrActionCreators);
       return reducerCreator;
     },
+
+    injectLogic(logicCreator) {
+      return logicCreator(reducerCreator);
+    },
   };
   return reducerCreator;
 }